Adding a Ping call to the session check loop
This code fixes an issue were tcp keepalive does not keep a connection
open above the ELB timeout, and tcp connections are dropped. The next
request against that session will fail, mark the session as unavailable,
and the session will reconnect for the next request, however a number of
requests do fail in this time (equal to the pool size).

An alternaitve to this change would be to work some retry logic into the
request cycle of the package to retry should the session be discovered to
be unavailable when the request fails.